backend-java/E-Commerce/catalogue-service/src/main/java/be/pxl/services/controller/CategoryController.java
package be.pxl.services.controller;

import be.pxl.services.domain.Category;
import be.pxl.services.service.ICategoryService;
import lombok.RequiredArgsConstructor;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;

@RestController
@RequestMapping("/api/categories")
@RequiredArgsConstructor
public class CategoryController {

    private final ICategoryService categoryService;

    @GetMapping("/")
    public List<CategoryResponse> getAllCategories() {
        return categoryService.getAllCategories();
    }

    @PostMapping("/createCategory")
    public ResponseEntity<Category> createCategory(@RequestBody Category newCategory){

        categoryService.createCategory(newCategory);

        return ResponseEntity.ok(newCategory);
    }
}
